lab4/requirements.txt
fastapi==0.70.0
uvicorn
SQLAlchemy==1.4.40
alembic
celery==5.2.3
redis==4.3.4
psycopg2-binary 
flower==1.2.0
pyjwt 
bcrypt
passlib
python-multipart
uvicorn[standard]